policy: move PolicyRepo & PolicyUpdater providers to policy cell
After removing the `newPolicyTrifecta` function with #34332, the
`PolicyRepo` & `PolicyUpdater` are still provided by functions
in the daemon package.

This commit moves them into their own cell in the package
`pkg/policy/cell`. The reason for the subspackage is to break
circular build dependencies.
