+++ daemon/cmd/policy.go
@@ -20,16 +20,12 @@ import (
 	. "github.com/cilium/cilium/api/v1/server/restapi/policy"
 	"github.com/cilium/cilium/pkg/api"
 	"github.com/cilium/cilium/pkg/clustermesh"
-	cmtypes "github.com/cilium/cilium/pkg/clustermesh/types"
-	"github.com/cilium/cilium/pkg/crypto/certificatemanager"
 	"github.com/cilium/cilium/pkg/endpoint"
 	"github.com/cilium/cilium/pkg/endpoint/regeneration"
 	"github.com/cilium/cilium/pkg/endpointmanager"
-	"github.com/cilium/cilium/pkg/envoy"
 	"github.com/cilium/cilium/pkg/eventqueue"
 	"github.com/cilium/cilium/pkg/identity"
 	"github.com/cilium/cilium/pkg/identity/cache"
-	"github.com/cilium/cilium/pkg/identity/identitymanager"
 	"github.com/cilium/cilium/pkg/ipcache"
 	ipcacheTypes "github.com/cilium/cilium/pkg/ipcache/types"
 	"github.com/cilium/cilium/pkg/k8s/synced"
@@ -45,73 +41,6 @@ import (
 	"github.com/cilium/cilium/pkg/source"
 	"github.com/cilium/cilium/pkg/time"
 )
-
-type policyRepoParams struct {
-	cell.In
-
-	Lifecycle       cell.Lifecycle
-	CertManager     certificatemanager.CertificateManager
-	SecretManager   certificatemanager.SecretManager
-	IdentityManager *identitymanager.IdentityManager
-	ClusterInfo     cmtypes.ClusterInfo
-}
-
-func newPolicyRepo(params policyRepoParams) *policy.Repository {
-	if option.Config.EnableWellKnownIdentities {
-		// Must be done before calling policy.NewPolicyRepository() below.
-		num := identity.InitWellKnownIdentities(option.Config, params.ClusterInfo)
-		metrics.Identity.WithLabelValues(identity.WellKnownIdentityType).Add(float64(num))
-		identity.WellKnown.ForEach(func(i *identity.Identity) {
-			for labelSource := range i.Labels.CollectSources() {
-				metrics.IdentityLabelSources.WithLabelValues(labelSource).Inc()
-			}
-		})
-	}
-
-	// policy repository: maintains list of active Rules and their subject
-	// security identities. Also constructs the SelectorCache, a precomputed
-	// cache of label selector -> identities for policy peers.
-	policyRepo := policy.NewStoppedPolicyRepository(
-		identity.ListReservedIdentities(), // Load SelectorCache with reserved identities
-		params.CertManager,
-		params.SecretManager,
-		params.IdentityManager,
-	)
-	policyRepo.SetEnvoyRulesFunc(envoy.GetEnvoyHTTPRules)
-
-	params.Lifecycle.Append(cell.Hook{
-		OnStart: func(hc cell.HookContext) error {
-			policyRepo.Start()
-			return nil
-		},
-	})
-
-	return policyRepo
-}
-
-type policyUpdaterParams struct {
-	cell.In
-
-	Lifecycle        cell.Lifecycle
-	PolicyRepository *policy.Repository
-	EndpointManager  endpointmanager.EndpointManager
-}
-
-func newPolicyUpdater(params policyUpdaterParams) *policy.Updater {
-	// policyUpdater: forces policy recalculation on all endpoints.
-	// Called for various events, such as named port changes
-	// or certain identity updates.
-	policyUpdater := policy.NewUpdater(params.PolicyRepository, params.EndpointManager)
-
-	params.Lifecycle.Append(cell.Hook{
-		OnStop: func(hc cell.HookContext) error {
-			policyUpdater.Shutdown()
-			return nil
-		},
-	})
-
-	return policyUpdater
-}
 
 type identityAllocatorParams struct {
 	cell.In
